backend/pyproject.toml
[tool.poetry]
name = "resonite_communities"
version = "0.1.0"
description = "Your project description here"
authors = ["brodokk <brodokk@brodokk.space>"]

[tool.poetry.dependencies]
python = "^3.10"
Flask = "^2.3.3"
requests = "^2.28.1"
gunicorn = "^21.2.0"
apscheduler = "^3.10.4"
toml = "^0.10.2"
oauth2client = "^4.1.3"
google-api-python-client = "^2.99.0"
google-auth-httplib2 = "^0.1.1"
google-auth-oauthlib = "^1.1.0"
python-dateutil = "^2.8.2"
timeago = "^1.0.16"
redis = "^5.0.0"
disnake = "^2.9.0"
easydict = "^1.10"
jsonschema = "^4.19.0"
Flask-Discord = { git = "https://github.com/brodokk/Flask-Discord.git", branch = "v10_custom" }
fenkeysmanagement = "^1.0.1"
resonitepy = { git = "https://github.com/brodokk/resonitepy.git" }
uvicorn = "^0.24.0.post1"

[tool.poetry.scripts]
web_client = "resonite_communities.server.web_client:run"
signals_manager = "resonite_communities.signals.manager:run"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"